새로 만든 팩토리와 시더에 문제 있음

+++ database/migrations/2022_02_23_004729_create_products_table.php
@@ -16,8 +16,8 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();                                          // 제품 고유 번호(PK)
             $table->string('productName')->nullable();     // 제품명
-            $table->smallInteger('price')->nullable();      // 제품 가격
-            $table->smallInteger('stock')->nullable();      // 제품 재고 수량
+            $table->unsignedBigInteger('price')->nullable();      // 제품 가격
+            $table->unsignedBigInteger('stock')->nullable();      // 제품 재고 수량
 
             $table->string('title')->nullable();             // 공동구매 글 제목
             $table->text('contents')->nullable();            //    -    글 내용
